Reallizado correcoes do crud

+++ back_end/Gestao_Escolar/src/main/java/com/example/GestaoEscolar/DTO/CadastroDto.java
@@ -1,6 +1,6 @@
 package com.example.GestaoEscolar.DTO;
 
-public record CadastroDto(String nomeCompleto, Integer ra, String endereco, String bairro,
-                          Integer cep, Integer telefone, String nomeMaeCompleto, String nomePaiCompleto,
+public record CadastroDto(String nomeCompleto, Long ra, String endereco, String bairro,
+                          Long cep, Long telefone, String nomeMaeCompleto, String nomePaiCompleto,
                           String usuario, String senha) {
 }
